sanic/requirements.txt
aerich==0.3.3
aiofiles==0.6.0
aiosqlite==0.16.0
asyncpg==0.21.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
click==7.1.2
flake8==3.8.4
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==2.10
iso8601==0.1.13
mccabe==0.6.1
multidict==5.0.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.2
pyflakes==2.2.0
PyJWT==1.7.1
PyPika==0.44.0
python-dotenv==0.15.0
rfc3986==1.4.0
sanic==20.9.1
sanic-jwt==1.5.0
six==1.15.0
sniffio==1.2.0
tortoise-orm==0.16.17
typing-extensions==3.7.4.3
ujson==4.0.1
uvloop==0.14.0
websockets==8.1
